requirements.txt
# Ultralytics requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
opencv-python>=4.6.0
Pillow>=7.1.2
PyYAML>=5.3.1
# requests>=2.23.0
requests==2.11.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0

# Add (sooin added) ---------------------------
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.5
fonttools==4.38.0
gcloud==0.17.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.3
mpmath==1.3.0
numpy==1.21.6
oauth2client==3.0.0
opencv-python==4.7.0.72
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
protobuf==4.23.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.4.3
pyparsing==3.0.9
Pyrebase==3.0.27
pytesseract==0.3.10
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-toolbelt==0.7.0
rsa==4.9
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.23.1
six==1.16.0
sympy==1.10.1
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
ultralytics==8.0.105
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0

# Logging -------------------------------------
# tensorboard>=2.13.0
# clearml
# comet

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=6.0  # CoreML export
# onnx>=1.12.0  # ONNX export
# onnxsim>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TF exports (-cpu, -aarch64, -macos)
# tflite-support
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev>=2022.3  # OpenVINO export

# Extras ———————————————————
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
# ipython  # interactive notebook
# albumentations>=1.0.3
# pycocotools>=2.0.6  # COCO mAP
# roboflow
Pyrebase~=3.0.27